# api/docs/swagger.yaml
basePath: /services
definitions:
...
  api_internal_routes_middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  api_pkg_utils.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  internal_services_oc_group.CreateGroupRequest:
    properties:
      config:
        $ref: '#/definitions/ocgroup.OcservGroupConfig'
      name:
        type: string
    required:
    - config
    - name
    type: object
...
  internal_services_panel.CreateSiteConfigRequest:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  internal_services_panel.GetFullPanelConfigResponse:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  internal_services_panel.GetPanelConfigResponse:
    properties:
      google_captcha_site_key:
        type: string
      init:
        type: boolean
    type: object
  internal_services_panel.UpdateSiteConfigRequest:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
...
  internal_services_user.CreateAdminUserRequest:
    properties:
      password:
        example: doe123456
        maxLength: 16
        minLength: 2
        type: string
      username:
        example: john_doe
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  internal_services_user.CreateAdminUserResponse:
    properties:
      token:
        type: string
    type: object
...
  ocgroup.OcservGroupConfig:
    properties:
      dns:
        items:
          type: string
        type: array
      dpd:
        type: integer
      idle-timeout:
        type: integer
      ipv4-network:
        type: string
      keepalive:
        type: integer
      max-same-clients:
        type: integer
      mobile-dpd:
        type: integer
      mobile-idle-timeout:
        type: integer
      mtu:
        type: integer
      no-udp:
        type: boolean
      restrict-user-to-routes:
        type: boolean
      rx-data-per-sec:
        type: string
      session-timeout:
        type: integer
      stats-report-time:
        type: integer
      tunnel-all-dns:
        type: boolean
      tx-data-per-sec:
        type: string
    type: object
  ocgroup.OcservGroupConfigInfo:
    properties:
      config:
        $ref: '#/definitions/ocgroup.OcservGroupConfig'
      name:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /api/v1/ocserv/groups:
    get:
      consumes:
      - application/json
      description: List Of Groups Sort By Name
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ocgroup.OcservGroupConfigInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: List Of Groups
      tags:
      - Ocserv Group
    post:
      consumes:
      - application/json
      description: Create Ocserv Group by given name
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: oc group config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_oc_group.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Create Ocserv Group
      tags:
      - Ocserv Group
  /api/v1/ocserv/groups/:name:
    delete:
      consumes:
      - application/json
      description: Delete Ocserv Group by given name
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Delete Ocserv Group
      tags:
      - Ocserv Group
    post:
      consumes:
      - application/json
      description: Update Ocserv Group
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group Name
        in: path
        name: name
        required: true
        type: string
      - description: oc group config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocgroup.OcservGroupConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Update Ocserv Group
      tags:
      - Ocserv Group
  /api/v1/ocserv/groups/defaults:
    post:
      consumes:
      - application/json
      description: Update Ocserv Defaults Group initializing step
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: oc group default config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocgroup.OcservGroupConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Update Ocserv Defaults Group
      tags:
      - Ocserv Group
  /api/v1/ocserv/groups/names:
    get:
      consumes:
      - application/json
      description: List Of Group Names Sort By Name
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: List Of Group Names
      tags:
      - Ocserv Group
...
  /api/v1/panel/config:
    get:
      consumes:
      - application/json
      description: Get Panel Config
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_panel.GetFullPanelConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Get Panel Config
      tags:
      - Panel
    post:
      consumes:
      - application/json
      description: Create Panel Config initializing step
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: check secret key from file 'init_secret'
        in: query
        name: init_secret
        required: true
        type: string
      - description: site config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_panel.CreateSiteConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Create Panel Config
      tags:
      - Panel
    put:
      consumes:
      - application/json
      description: Update Panel Config in app after login step
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: site config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_panel.UpdateSiteConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Update Panel Config
      tags:
      - Panel
  /api/v1/panel/config/init:
    get:
      consumes:
      - application/json
      description: Get Panel Config to discover init data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_panel.GetPanelConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
      summary: Get Panel Config with Init
      tags:
      - Panel
...
  /api/v1/user/admin:
    post:
      consumes:
      - application/json
      description: Create Superuser in initializing step
      parameters:
      - description: check secret key from file 'init_secret'
        in: query
        name: secret
        required: true
        type: string
      - description: admin user body data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_user.CreateAdminUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_user.CreateAdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
      summary: Create Superuser
      tags:
      - Site User
...
swagger: "2.0"